requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
environ==1.0
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.8
Werkzeug==3.0.0
